Module bolero::generator

Re-export of bolero_generator

Re-exports

pub use bolero_generator;

Traits

OneOfExt

Extensions for picking a generator from a set of generators

OneValueOfExt

Extensions for picking a value from a set of values

TypeGenerator

Generate a value for a given type

TypeGeneratorWithParams

Convert a type generator into the default value generator

ValueGenerator

Generate a value with a parameterized generator

Functions

constant

Always generate the same value

gen

Generate a value for a given type

gen_with

Generate a value for a given type with additional constraints

one_of

Pick a generator for the provided set of generators

one_value_of

Pick a value for the provided set of values

Derive Macros

TypeGenerator

Derive the an implementation of TypeGenerator for the given type.